requirements.txt
requests
flask
nintendoclients
